requirements.txt
Django==4.2.5
Django==5.0.4
djangorestframework==3.15.1
google_api_python_client==2.127.0
google_auth_oauthlib==1.2.0
protobuf==5.27.1
python-dotenv==1.0.1
